timeless/requirements.txt
APScheduler==3.9.1
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.3